requirements.txt
requests
ccxt
six
flask
python-dateutil
matplotlib
pandas